Clean up params to private methods in push command
This command has been worked on a lot in the last week and we didn't
always keep a close eye on whether or not the parameters to private
functions still made sense. This commit remove some unused params
and inlines a function call that was only used in a single place,
and was only a single line.