Make tests pass on CI again
On both of our windows ci machines, these tests fail when the websocket
handlers don't sleep before closing the channel. This test is fairly
roundtrip, so it appears that the issue is that we're having trouble
making these blocking reads. Even setting the size of the channel we
read from to zero (as @tedsuo suggested at one point, I believe) has
no effect.