Remove fixture
This breaks out windows CI machines, because golang does not recognize a
unix symbolic link file as having a special mode. This isn't very
surprising because go's FileMode is intended to be a platform agnostic
way of asking "does this file have special bits set (for my platform)?".

This would be best tested with a platform specific integration test.