'cf create-service' is more backwards compatible
[Finishes #67141076]

We don't have this query by label available on earlier releases that we need to
support, so we need to be able to fall back and find the matching services. It's
probably fine to implement this by looking at the error code because it's not clear
when CC has a release with the api change.