Cleanup delete-app tests
Also add an explicit login requirement. We worked around this at
some point by making the gateway aware of requiring the user to login
when the refresh token is invalid. Not the best solution because of the
mising of concerns, but taking these steps now will make that code easier
to clean up in the future.

[#66911426]