Update README instructions for checking out the repo
My hope is that explaining how to get started with the repo will make it easier for
other contributors. These steps assume some basic knowledge of git and GO, but the
only way we will improve this documentation is to start improving it today and continue
to be sympathetic to pain the community feels in the future.

[Finishes #70686778]