Fix double confirm prompt bug
This was a regression introduced ... a few iterations ago. Thinking
about how these tests are constructed now, I wonder if our hesitance to
write tests that are directly tied to the implementation (e.g.: fragile
tests that assert on the actual string value) is misguided. We would
never have caught this bug in our unit tests.

[Finishes #70960378]