Remove unneeded godep from build
As described in this GitHub comment[1] the use of godep in the build
script is unneeded.  Having it there requires all builders of the
project to have it installed even though it never needs to be called.
As an example, this causes the Homebrew formula for the project to be
massively more complex than it needs to be.  This change removes that
use of godep in the buildscript, instead calculating the absolute path
to the checked-in godep workspace and using that.

[1]: https://github.com/XenoPhex/homebrew/commit/a885c669908886d4cf9214ed1e812bd29dc7e217#commitcomment-6494522