Improve usage of codegangsta/cli
Recently the help printer interface for codegangsta/cli was updated to
also be called for command help. That's pretty useful but we were
relying on it always for printing JUST the app help.

This change actually keeps our specially templated app help but also
prevents codegangsta from calling our app help printer for commands.

Basically, this is still just a giant ticking bomb.