Unclog all the pipes
We discovered this problem because our builds on windows reached the
maximum amount of time (10 minutes) in the cf/app package. Narrowed it
down to executing the template and our theory was that the amount of
characters being written overall in the template to stdout is what
caused this to block indefinitely. Under test, we swap out stdout and
stdin and expect to be able to just keep writing indefinitely. That's
probably not a valid assumption. Probably.

[#73899232]