Move api-switching logic into Requirements phase
There are two flags that need to be checked for; set_roles_by_username
and unset_roles_by_username.

This necessitates moving the logic out of the UserRequirement and back
into the command, but higher up than it was before.

Presently UnsetSpaceRole passes 'true' always in the call to
NewUserRequirement. This is in accordance with how it worked previously
and will be updated in a future story.

[#97841328]

Signed-off-by: Simon Leung <leungs@us.ibm.com>
Signed-off-by: Kris Hicks <khicks@pivotal.io>