Make date reported in --version semver compliant
Remove ':' and '+' after the first plus, per semver.org:

    Build metadata MAY be denoted by appending a plus sign and a series
    of dot separated identifiers immediately following the patch or
    pre-release version. Identifiers MUST comprise only ASCII
    alphanumerics and hyphen [0-9A-Za-z-]. Identifiers MUST NOT be
    empty.  Build metadata SHOULD be ignored when determining version
    precedence.  Thus two versions that differ only in the build
    metadata, have the same precedence. Examples: 1.0.0-alpha+001,
    1.0.0+20130313144700, 1.0.0-beta+exp.sha.5114f85.

It's good to note that this isn't actually part of the version when we
do version comparisons within the code, but we present it as if it's
part of the version in `cf --version`, so it should look correct.

Later we'll use the semver resource in Concourse to produce proper
semver versions, at which point we *will* include it in the comparison.