Fix RouteCreator interface
This wasn't being properly tested previously, and the type assertion
that is done in map-route isn't checked anyway, which meant this
resulted in an interface conversion panic.

This particular interface (route.RouteCreator) is being tested directly
in create_route_test, which means we can add a type assertion before
running those tests. In tests for other commands, however, we test only
the Execute() method. In those cases we should stop doing that and test
the interface method directly in addition to testing the Execute()
method to show that Execute() is actually calling the interface method,
or by at the very least doing the type assertion in a test to ensure
that changes to the method don't stop the command from being an
implementer of the interface.

[#100992242]