properly handle '.*' in .cfignore
We were skipping the entire app directory because '.*' matches the
unexpanded '.' (this directory). When walking the directory we now check
the full path against the app directory path first and continue if it
matches, instead of skipping the entire directory. The test just makes
sure that the app directory is not skipped (it still gets other files to
upload), but dotfiles are ignored when '.*' is present in the cfignore.

(github issue #870) [fix #122234577]
Signed-off-by: Nick Wei <nwei@pivotal.io>