Update install instructions for binaries
During user testing for installing the CLI we discovered that people had issues using the
curl based install. The binary is not on the path, and the instructions generally in the documentation
assume it is. For example cf --version rather than ./cf --version.

By suggesting moving to somewhere in your path like a common location like /usr/local/bin the confusion
can be reduced.