UI no longer merges map, config is written to disk on command exit
- The translate function that i18n provides does not merge the keys into
the template string if it cannot find the template string in the
resource files. A wrapper was written around it to merge the keys into
the template even when the template string isn't found.
- The map merging functionality was removed from UI as it is probably
useless in the long-term and too much of a burden to deal with when
modifing a UI function.
- Config files are now written to disk after a command completes
execution.
- Added API related config commands

[#126256625]

Signed-off-by: Nick Wei <nwei@pivotal.io>