Reset timeout in callback
The issue with the previous code was that data had to be received over the connection
for the reconnect to be considered valid.  The new logic makes use of the OnConnectCallback.

The test for the onConnect was explicitly asserting that OnConnectCallback was called
with the onConnect function directly, rather than asserting that calling the
OnConnectCallback's argument would call onConnect.  We updated the test to be a bit
less explicit about that, but still test the desired behavior.

[#131582719]

Signed-off-by: Sam Nelson <snelson@pivotal.io>