mention how to get invited to CF slack in CONTRIBUTING doc
The CONTRIBUTING document recommends checking in with the team before
starting work on a pull request either via a Github issue or on the #cli
channel on CF slack.

However, first-time contributors may be unaware of the process for getting an
invite to the Cloud Foundry slack team. This commit adds a breadcrumb
for people to follow.