copy cats into gopath instead of using path: in task yaml
Using path: makes the cf-acceptance-tests in the gopath a symlink on
Concourse 3.0.1. When running ginkgo with multiple nodes on Windows
some of the nodes will randomly fail with:

chdir .: The process cannot access the file because it is being
used by another process.

Running the same binary on multiple ginkgo nodes does not seem to work
if the test cmd.Dir is a symlink.