On an exit signal, wait for plugins to exit
When a user running a plugin presses `CTRL+C` the CF CLI and the
plugin will receive a `SIGINT` signal. Until now, the CF CLI and its
RPC server exit immediately--despite that the plugin may have cleanup
to do.

In this commit we have the CF CLI ignore common exit signals. If a
keyboard interrupt has been used then the plugin will receive the
signal and exit itself.

The only situation not handled is where *only* the CF CLI is
signalled, as the signal will not be relayed to the plugin. We are
uncertain how to resolve that and believe it can be consider later.