Fix query parameter regex in trace.Sanitize()
The previous regular expression, &password=[^&}*&, did not match password
parameters at the end of the query string. In that case it also ate the rest of
the input, until the next & character.

I replaced it with an expression that is, sadly, rather complicated, but should
match all characters that can be part of a query parameter value.