add steps for poll package through restart app
Note: There are some interesting issues with the logger method in
`v3/v3_push_command.go`. The logger is a halting operation the requires
the CLI to at some point close the NOAA client. We're currently stopping
it after `StagingComplete`. However, this leads to some possible race
conditions regarding the logger and the UI. We need more investigation
in regards to how v2 closes the NOAA client connection.

[#157577905]

Signed-off-by: Sebastian Vidrio <svidrio@pivotal.io>
Signed-off-by: An Yu <anyu@pivotal.io>