display v3 app display in v2 push
Display the start command only when pushing - a lot of code had to be
added to optionally retrieve and display the start command. This is due
to several reasons:
- When listing a process - the command is redacted by the API. This
means an additional get is required [per process] in order to get the
command.
- The additional call is not required in all but the push case (push
displays the 'start command' field), so all of these other requests
should not make these extra calls.
- The 'optional' nature of these calls requires passing
'withObfuscatedValues' bool to every place process summaries are used in
order to skip the additional calls.

[Finishes #158607105]