Refactor create-buildpack error handling
In the create-buildpack command, there are two ways to get a
name collision. First, there is a buildpack with the name you are trying
to use, that has a nil stack name. Second, there is a buildpack that
has the name you are trying to take, and the manifest that is uploaded
with created buildpack specifies the same stack.

Previously, it was not obvious why the code needed to check for a name
collision after uploading the bits of the buildpack. This refactor
is to make it more obvious why there are two checks.

[#126256865]

Co-authored-by: Tom Viehman <tviehman@pivotal.io>