Stop matching on external error message text
In create buildpack and update buildpack name collision errors,
the copy that is eventually printed to the user is returned from CAPI.
Because this team does not own the text of that error message, we want
the integration tests to stop failing when the copy is different than
what we expect. We still assume the message will contain the name of
the buildpack that collided.

[#158614276]

Co-authored-by: Tom Viehman <tviehman@pivotal.io>