Fix error message on UAA version too low
The error message displayed when passing `--origin` was confusing
because it sounded like it was reporting that the version of CAPI was
too low, but it was actually reporting that the version of UAA was
too low.

We added a way for integration test helpers to check the version of
UAA so that the tests know whether they should run against a given
environment.

Also, several tests around `--origin` were incorrect. We updated the
unit tests of the translateable error to test the correct error
type.

[Finishes #159781083]

Co-authored-by: Spencer Hawley <shawley@pivotal.io>