Warn user and exit if buildpack path is empty directory
For the create and update buildpack commands, the user can specify
a path to a directory to be zipped up and uploaded as a buildpack.
Previously, if the user passed the path to an empty directory, the CLI
would upload a zip file representing the empty directory, which would
result in an unusable buildpack. Now the CLI will inform the user
that empty directories are not allowed and exit 1 when this happens.

[Finishes #159501566, #159532186]

Co-authored-by: Spencer Hawley <shawley@pivotal.io>