Display service binding operation type
The cf service <service_name> command displays a table of bound
applications and allows users to discover information about bindings.
For asynchronous bindings this includes a state, type and description.
This commit adds the type field.

It is now displayed under the status column by combining the state and
type. E.g. type: 'create' and 'state': 'in progres' becomes the
status string 'create in progress'. This is consistent with displaying
asynchronous service instance information.

[#160024565]

Signed-off-by: Alex Blease <ablease@pivotal.io>
Co-authored-by: Alex Blease <ablease@pivotal.io>