Put stacks last in output of "cf buildpacks"
For backwards compatibility with scripts that may be parsing the output
of `cf buildpacks`, we moved the new column "stacks" to the rightmost
position in the output.

Also add integration tests for buildpacks command, and add a helper
method for checking the unrefactored commands' behavior when the user
is not logged in or not targeting a CF correctly. (Old commands write
these errors to STDOUT instead of STDERR.)

[Finishes #160187405]

Co-authored-by: Ryan Tang <rtang@pivotal.io>