Zero Downtime Push needs CAPI 3.57.0
Explicitly fail with a sensible error if the experimental
zero downtime push command is run against a CAPI too old to support it.

CAPI must return "deployments" in the list of resources return by "/v3"
or the code will not work, but will fail in a confusing way. This checks
for the first version where we expect "/v3" to behave correctly and
gives a sensible error message if the version is not met.

[#160008817]

Co-authored-by: Will Murphy <wmurphy@pivotal.io>