improve bosh-lite script and contributing docs
* Make the "clean" functionality more reliable
* Switch the CF to using postgres due to a bug in the MySQL release that
caused the MySQL proxy to lose its connection to the database, which in
turn caused the `cloud_controller_ng` job to fail.
* All assets (vars-stores, state files, manifest) created by the
bosh-lite script will be stored in the same directory, which defaults to
~/workspace/cli-lite.
* The bosh-lite script will store a copy of the CF manifest (without any
variables interpolated) in the cli-lite directory; this is to make
iterating on changes to the script easier and less reliant on re-running
the entire script with areas commented out.
* Refactor the script for readability.