rename AppCommand -> V3AppCommand
This command will move from the V3 (aka CLI V7) folder to the CLI
V6 folder requiring that either this command be renamed or the original
v2 version be renamed. Since there is limited V3 in the V6 CLI, we're
opting to be more explicit about the V3 version of app.

[#160853831]