Build OSX binary on Darwin and allow cgo DNS
We've had reports of DNS issues on OSX caused by known incompatibility
between OSX DNS functionality and the Golang net package's dns resolver.
The current way to build binaries that are compatible with OSX DNS is to
build them with cgo enabled and to not force use of the net package dns
resolver (so, remove the netgo build tags). In order to build the OSX
binary with minimal complexity, we have to stop cross-compiling.

There is a risk that this change will re-introduce an issue with
self-signed certificate chains that we addressed in
[this tracker bug](https://www.pivotaltracker.com/story/show/156851937)
by switching to building statically-linked Darwin binaries.

[#159627921]

Co-authored-by: Tom Viehman <tviehman@pivotal.io>