Fix host and path combining in paginated requests
Previously, paginated requests generated the "next page" url by
concatenating the API url with the next page path returned by CAPI. In
cases where the API url was specified with a trailing path, this could
result in URLs like "https://api.example.com//v2/organizations". Older
versions of CAPI would return a 404 in response to such a URL.

Now we use go's url package to append the path, so that edge cases such
as double slashes will be handled by a library.

We also added a trailing slash to the API configured in the unit tests.

[Finishes #161006428]

Co-authored-by: Will Murphy <wmurphy@pivotal.io>