Allow create-space to run with client credentials
UAA clients do not have a username associated with them, so we need to
use the GUID-based role endpoints in CloudController in order to grant
clients the necessary permissions. We also need to make the client an
org member before assigning space permissions, otherwise the
CloudController will fail with invalid relation errors.

[#159970624]

Co-authored-by: Tom Viehman <tviehman@pivotal.io>