Use CC extra structure for documentation urls
In past, the CC used to provide the now deprecated `documentation_url`
field in the entity structure of a service instance. At some point, to
align with the OSB API spec's concept of provider metadata, we moved
this kind of information under the `extra` structure and the previous
field started returning nulls. This broke the `cf service
<service_instance>` command, which started providing empty documentation
urls.

This introduces a backwards compatible fix for this behaviour,
supporting both the `documentation_url` under `entity` and
`documentationUrl` under `extra`. There should never be a conflict
between these two (and no CC should provide both since the change was breaking),
but to be careful, we will always favour the old behaviour.

[finishes #157060329]

Co-authored-by: Henry Stanley <hstanley@pivotal.io>