Updated NullInt to be compatible with go-flags changes
The fork of go-flags that we use now expects a flag that requires custom
value validation to define an IsValidValue method that handles
validating that flag's inputs. These changes to the fork are being done
so that we can eventually move off of the fork and back to the main
go-flags repository.

[#160742110]

Co-authored-by: Tom Viehman <tviehman@pivotal.io>