Add gofmt check for integration tests
Because the "cli" input to our units job ignores the integration folder,
it will not pick up commits that fix gofmt errors if the commits in
question have no changes to unignored paths. In this scenario, the units
job will not pass until a commit modifying unignored paths is pushed. To
remedy this, we will check that the integration folder has been go
formatted as part of the integration job(s).