switch go-flags to rely on 'cli-vendor' branch
We are currently attempting to land a pull-request in the upstream
go-flags repository. This PR would allow us to abandon our own forked
version of go-flags.

However, we want to be able to make these attempts separate from having
to apply each one of them to the CLI codebase.

By creating a 'cli-vendor' branch, the CLI can rely on a stable fork as
opposed to this moving PR.

[#160742110]

Signed-off-by: Anand Gaitonde <agaitonde@pivotal.io>