Refactor create-shared-domain
The bulk of the code change is adding a router client, since the
create-shared-domain command needs to be able to query the routing API.

The refactor also includes improvement for error handling.

For now, the routing API client largely duplicates the CloudController
and UAA clients. We may extract this shared code later.

[Finishes #126256725]

Co-authored-by: Will Murphy <wmurphy@pivotal.io>
Co-authored-by: Tom Viehman <tviehman@pivotal.io>
Co-authored-by: Magesh Kumar Murali <mmurali@pivotal.io>