Refactor marketplace command
This commit is split into a few parts.

Firstly, it introduces some utility methods on the SharedActor for
checking whether the CLI is logged in, targeting a space and targeting
an org. The CheckTarget and Require methods has been refactored to make
use of this new functionality. These new methods were abstracted over
the very similar methods on the Config, but we thought that it would be
better to provide a layer of indirection to commands, such that we could
later decide to determine these behaviours in a different manner.

Secondly, we introduced a set of integration tests for the marketplace
command. They live in the global integration suite as they require
registration of service brokers, and enabling access to those services,
both of which result in global visibility changes that could affect
other tests.

Finally, we moved the marketplace command to the new v2 style,
introducing actor and cloud controller client methods to look up all
services, and a specfic service by name from both the global services
endpoint and by space.

Signed-off-by: Luis Urraca <lurraca@pivotal.io>
Signed-off-by: Sam Gunaratne <sgunaratne@pivotal.io>
Signed-off-by: Georgi Lozev <georgi.lozev@sap.com>
Signed-off-by: Alex Blease <ablease@pivotal.io>
Signed-off-by: Jen Spinney <jspinney@pivotal.io>
Signed-off-by: Henry Stanley <hstanley@pivotal.io>

[finishes #160686306]
[PR]