Finished create-shared-domain refactor
Refactored version of create-shared-domain is now the default.
Added error wrapper and better testing to router client.
Cleaned up some duplicated code that was introduced by
6b5c4a4addbfaac44ef005eea964fa30fdcafb91.

We decided not to add a retry wrapper to the rouer client at this time
because it would have to be duplicated from an existing client and we
only call the routing API in one situation. Hopefully future refactors
will reduce the need to duplicate wrappers for each client.

[Finishes #126256725]

Co-authored-by: Magesh Kumar Murali <mmurali@pivotal.io>
Co-authored-by: Tom Viehman <tviehman@pivotal.io>
Co-authored-by: Jen Spinney <jspinney@pivotal.io>