Added Unauthorized Error to the router client
When the router client receives unauthorized message "You are not authorized to
perform the requested action", we used to send RawHTTPStatusError
object. This caused the UI to print the raw message which was not the
desired behavior. This commit creates a new error object
(UnauthorizedError) for this
scenario and returns this custom object which the actor and command
layer can use.

[finishes #126256725]