Fix incorrect test assertion
In package_test, we have been passing the wrong URL and method to a
VerifyRequest. This hasn't been a problem in the past because
of the "luck" of goroutine ordering, which resulted in that
verification never actually running.

Now we are hitting VerifyRequest and need to have the correct URL, etc.
in the test assertion.

We also reduced the scope of a variable in an another context to match the scope it's used in.

[#160911621]

Signed-off-by: Will Murphy <wmurphy@pivotal.io>