Remove -race from windows units temporarily on V7 side.
Because of golang/go#27089, passing "-race" to
ginkgo causes the windows build to fail. We are simply removing this
flag until the golang issue is resolved. (Units on other operating
systems still check for race conditions.)

This change is to parallel fe63a2641cd035706f691ad07594cb42a94bf5bc.

[#160911621]