Change v3 Process to use FilteredString
An app dev should be able to use the `-c` flag and provide a string of
`null` or `default` to reset the start command of their app to the
detected/default start command. We need to send a `null` value to the
CCv3 API. We prefer to represent this as a FilteredString rather than a
null pointer.

[Finishes #161797187]

Signed-off-by: Mike Kenyon <mkenyon@pivotal.io>