Fix racy ghttp server in ccv2 unit tests
Previously, there was a single server being set up in a
SynchronizedBeforeSuite block to be shared between all tests, with an
attempt to clean state via `server.Reset()`. However, when the tests are
run in parallel, it is still possible to have a sequence where server
verification assertions interleave with incoming http requests from
concurrent tests.

The cost of starting a server appears to be negligble, so this commit
ensures there is an entirely new server set up for each spec that runs.