Increase default NODES for integration tests
In order to speed up the integration test runs we want to
parallelize more. If local runs of integration tests against bosh-lites
are flaky, we should decrease NODES further. We also changed the units
to use `-p` instead of `-nodes` for parallelization, which means that
ginkgo will automatically choose how many nodes to spin up.

[Finishes #160998631]