Apply temporary fix for flakey units
We've been seeing a lot of flakes in the ccv2/ccv3 suites that seem to
stem from our use of the ghttp Server in parallelized tests. In order to
keep our pipelines flowing while we continue the investigation, we're
changing the test setup to spin up a new ghttp Server for each test.
This change is identical to PR #1511.

[#162268129]

Co-authored-by: Jen Spinney <jspinney@pivotal.io>