Add workaround for --no-route
Routes were being created even though the manifest specified no route.
Because knowing not to create a route would require the CLI to parse the
manifest, we added a workaround to not create any routes when a manifest
is specified. This can cause apps to not get routes if routing
information is not specified in the manifest.

[Finishes #160358087]

Signed-off-by: Spencer Hawley <shawley@pivotal.io>