Add integration test job for new legacy environment
* In 2019, the new legacy environment (lola) will replace to the existing
legacy environment (homeroom).  This commit adds a new job to the
concourse pipeline to run integration tests in the new legacy
environment.
* The new legacy environment requires version 5 of the BBL CLI to
extract the correct values for connecting to credhub. This version of
BBL has been added to the cli-base image as bbl5 and is now referenced
in the cleanup-integration.yml concourse task.

[#162463658]