Ignore 404 on instance fetch for cf services cmd
Prior to this commit, there was a race condition, that service instances
could be listed, and service bindings for that instance could be listed,
but between the two requests, the service instance could be deleted and
the cf services command would fail.

This commit results in logging the error, but continuing with the rest
of the command.

This commit only applies to both managed and user provided service instances.

[finishes #162283179]
[PR]

Co-authored-by: William Martin <wmartin@pivotal.io>
Signed-off-by: William Martin <wmartin@pivotal.io>