Alert PM before redeploying Arsicault
We don't want to roll the acceptance env if Abby is in the middle of
doing acceptance. Applying an ops-file to the base infrastructure
pipeline allows us to continue using the base infra pipeline (which
ensures that any changes we make to our integration envs will also
happen to Arsicault).

Also refactors the reconfigure-pipelines script so that the function for
configuring infrastructure pipelines can be reused for environments with
non-standard pipeline configs (e.g. Lola and Arsicault)

[Finishes #162999668]