cf services command uses space summary endpoint
cf services command was much slower after the refactoring due to
making individual API calls to `/v2/services`, `/v2/service_plans` and
etc. For more details of the problem have a look at issue #1353

As part of this changes we switch back to using `/spaces/summary`
endpoint. It should significantly decrease the number of calls and thus
improve the performance. While working on the story we also
noticed there was a side effect introduced as part of the refactoring.
It was related to displaying bound applications from another space where your
service instance was shared to. We decided to stick with the old/legacy
behaviour. For more context on that have a look in the story
https://www.pivotaltracker.com/story/show/163052467 comments.
In this implementation we use the legacy approach to decide if a
service instance is user-provided or not, which is not ideal, but it
come at no cost.

[finishes #163052467]
[PR]

Co-authored-by: Alex Blease <ablease@pivotal.io>
Co-authored-by: Henry Stanley <hstanley@pivotal.io>
Co-authored-by: Aarti Kriplani <akriplani@pivotal.io>
Co-authored-by: Niki Maslarski <nikolay.maslarski@sap.com>
Signed-off-by: Alex Blease <ablease@pivotal.io>
Signed-off-by: Niki Maslarski <nikolay.maslarski@sap.com>
Co-authored-by: Niki Maslarski <nikolay.maslarski@sap.com>