Changed the mem_quota field to int64
This commit is related to Github issue (#1370). Previously, the
memo_quota field only accept a 32bit integer (when built on a 32 bit
machine). If we push an app with a memory quota more than 2147483647
bytes (2GB) then it would be unable to parse the json and return a json
unmarshal error.

This commit explicitly makes the mem_quota field an int64 type and hence
would fix the issue.

[#157074497]

Signed-off-by: Alexander Shan <ashan@pivotal.io>