Add -b flag to enable-service-access command
Since the CC has recently changed to allow the same broker to be
registered multiple times, it is possible to there to be naming
collisions on service offerings.

The -b flag allows the CLI to search for a service by name within the
context of a particular broker, allowing for disambiguation.

[finishes #162583614]

Signed-off-by: Alex Blease <ablease@pivotal.io>
Signed-off-by: William Martin <wmartin@pivotal.io>
Co-authored-by: William Martin <wmartin@pivotal.io>