Time out if claiming env lock takes too long
We've seen a few builds hanging forever because they can't claim an
environment lock. This situtation is very difficult to detect just from
looking at the pipeline - you have to click into the build itself, see
that the lock is unclaimed and then determine what the status of the
lock is.

Trying a 2hr timeout to start - if this causes any sort of false
positives, increase the timeout as necessary.