Adds integration tests covering SSO for cf login
The SSO flow for obtaining a temporary passcode is very browser-based.
We haven't yet found a non-brittle way to get a temporary passcode from
UAA's `/passcode` endpoint, so we are not actually testing the happy
path of the user entering a correct passcode.

Also, due to how the password prompting is implemented for windows, we
encountered difficulty asserting on the SSO passcode prompt (the prompt
did not appear on windows). Since this prompt doesn't come from our code
anyway (it originates from the UAA server), we are not testing the
prompts themselves.

[#163100310, #163069334]

Signed-off-by: Tom Viehman <tviehman@pivotal.io>