Unwrap three new buildpack upload errors in v7
Looking at the CAPI code base, the upload job can fail for these three
reasons that we were not accounting for. By unwrapping the actual error
from the JobFailedError, we improve the output shown to the user.

We stopped unwrapping 290001 because the v3 API never sends a
BuildpackNameTakenError. This was an error from back before buildpacks
had stacks.

[#162358208]

Signed-off-by: Mike Kenyon <mkenyon@pivotal.io>