automate generating of release note download URLs
Because V7 Beta releases should go on a wiki page rather than a true
github release, we can't use the github release resource. Instead, we're
using the github repo that backs the cli's wiki. Conveniently, the same
deploy keys that work for the github repo work for its cli repo.

[#161627605]

Signed-off-by: Leah Hanson <lhanson@pivotal.io>