Make integration tests work with SSL validation
* Need to provide the gorouter CA cert as well because communication to
Loggregator does not go through the GCP HTTPS load balancer; instead it
goes through a TCP load balancer, so TLS is terminating at the gorouter.
* Fixes tests that were unintentionally dependent on SSLDisabled being
set in the CF config file - when the boolean value of SSLDisabled
matches the value (presence/absence) of the --skip-ssl-validation flag
on `cf api` (and the URL/hostname provided is the same as the config
file), the command will no-op instead of actually hitting the API
endpoint.

[#163233217]

Signed-off-by: Alex Shan <ashan@pivotal.io>