Finish refactor of service-access command
This commit finishes the work started last year on to refactor the service-access command.

It's a fairly hefty set of changes, and the logic is quite tricky. We've tried our best to simplify
difficult cases but ultimately, the combinatorial nature of the -o, -b, -e flags is difficult.
There are a few changes that don't necessarily count as "refactoring", but are needed to fix frankly
confusing (and difficult to replicate) behaviours in the legacy command.

[finishes #162440699]

Signed-off-by: Niki Maslarski <nikolay.maslarski@sap.com>
Co-authored-by: Niki Maslarski <nikolay.maslarski@sap.com>
Signed-off-by: Aarti Kriplani <akriplani@pivotal.io>
Co-authored-by: Aarti Kriplani <akriplani@pivotal.io>