CLI user can use --vars-file for variable interpolation with manifests
Before, cf push would just read the manifest file and pass it through to
Conceptualize. Now, cf push uses the same ReadAndInterpolateRawManifest
that is used by v6. (We refactored ReadAndInterpolateManifest to pull
out the shared code. v6 wants a list of Application structs to return,
but v7 just wants a []byte.)

To test this new functionality, we added a simple unit test and three
integration tests (one vars-file, two vars-files, missing variable). In
adding the unit tests, we realized that the test manifest for the
existing unit tests was invalid (it had a literal \n rather than the
intended newline), so we fixed that. The new interpolation code will
Unmarshal and then Marshal all manifests, which creates non-meaningful
formatting changes. In order to avoid flakes if this formatting changes,
we added a function to Unmarshal and then Marshal all the expected
output strings.

[Finishes #162623888]

Signed-off-by: Eli Wrenn <ewrenn@pivotal.io>