Use 64-bit integers for deserializing server responses
Users on 32-bit systems reported issues where the CLI would fail to
unmarshal responses from the server because the response contained
integer values that, when unmarshalled, would overflow 32-bit integers.
This commit changes all integers that we receive from the CC to
explicitly use 64 bit types.

We also changed the Uptime field of ccv3's ProcessInstance to a
time.Duration because it allowed us to isolate knowledge of
converting the uptime into a duration to the ProcessInstance's
UnmarshalJSON method.

[#163397517]
Signed-off-by: Magesh Kumar Murali <mmurali@pivotal.io>
Signed-off-by: Tom Viehman <tviehman@pivotal.io>