Remove empty arguments from integration helper
V7 has a global switch that causes any additional arguments to fail.
In a previous commit, to dynamically toggle `--skip-ssl-validation`
in integration tests, we sometimes now pass in an empty argument to
the `cf api` command used for test setup.  The V7 global validation
still recognizes that empty string as an argument, which caused test
failures.

[finishes #163833779]

Signed-off-by: Jen Spinney <jspinney@pivotal.io>