Conditionally pass arg to skip ssl validation in tests
Previously, there was a suite level variable called `skipSSLValidation`
that set either to the empty string or to the string
"--skip-ssl-validation", and then this variable could be passed to gexec
sessions directly. But newer code paths fail with an unexpected argument
error if passed an empty string.

Therefore, change the suite level variable to a bool, and conditionally
pass either the string "--skip-ssl-validation" or nothing to gexec
sessions.

[#163886355]