Require '--endpoint' when 'cf push -u=http'
We want to require the user to specify the new HTTP health check
endpoint, rather than assume that it is '/' by default. Our default case
does not play well with v3 CAPI (due to how params get deeply-merged in
rails), so, we will make the user re-specify
the endpoint each time they push.

This code does not actually do anything with the new Endpoint. It just
requires that the user specify it. Stay tuned.

[#163804102]

Signed-off-by: Mike Kenyon <mkenyon@pivotal.io>