Remove Global marketplace test
These tests asserting there are no services is inherently flakey.
We've see it fail for a number of different reasons. The effort to
maintain it is higher than the value of having this as an end-to-end test.
We're testing the message that shows up when the array returned is empty,
which is adequadetly covered in unit tests too.

[finishes #163785631]

Co-authored-by: Aarti Kriplani <akriplani@pivotal.io>