Integration test: set-org-role with invalid org
This adds an integration test to validate that we display an error
message when the set-org-role command is given a non-existant org.

This commit also modifies the set_org_role_command_test file to use the
ReadOnlyOrg instead of creating a new org for each test. This is to
speed up test execution and cleanup.

[finishes #164089886]