Integration tests: set-space-role with invalid space/org
This commit adds two integration tests for the set-space-role command:
one test verifying the behavior of the command when given a non-existant
space, and one test verifying the behavior when given a non-existant
org.

This change also switches the tests to use the ReadOnlyOrg and
ReadOnlySpace to make test execution and cleanup run faster.

[finishes #164023518]
[finishes #164181946]