Add test for set_roles_by_username with invalid user
Since this test disables a feature flag (which means it modifies global
server state) it could pollute other tests that rely on the flag being
enabled, so it needs to be run in the global suite. The other tests for
set-org-role don't have the same restrictions so they can remain in a
parallelized suite.

[Finishes #164090104]