Warn user when running under-rewrite login command
This commit adds a new environment variable, CF_EXPERIMENTAL_LOGIN, that
we use to check whether to run the legacy or new login command. It
should be removed when the rewrite of login is concluded.

Changes also include a refactoring of the LoadConfig and Config env
tests to decouple the two areas. Previously, LoadConfig tests had been
through-testing methods on the Config object, and the Config env tests
had been through-testing LoadConfig's implementation which was creating
cumbersome coupling between the two implementations.

Note that the refactor does not include a full backfill of all tests
that may be missing from env_test.

[#164223388]