Refactor the Application UnmarshalYAML
We refactored this to allow the CLI maintainers to easily add fields
that the CLI need to know about to the ApplicationModel struct and then
access these from the parser's applications array as before.

Added the path and docker fields and modifed the tests to make sure that
the returned manifest for a specific app also contains those fields

Docker was left as a pointer in order to make the empty check easier.

[#163742342]

Signed-off-by: Connor Braa <cbraa@pivotal.io>