Fix test helpers for in memory server
In order to use a ghttp server as a testing fake in our integration
tests, we need to be able to map paths and queries to particular
handlers, but ghttp.Server can only map one handler per path. This adds
a test helper that registers stateful handlers with the underlying ghttp
server so that we can achieve these more detailed fakes.

As a proof-of-concept, this commit also adds a test for logging in as a
user with more than 50 orgs; the test uses the new fake server with
response fixtures generated from the real output of `cf orgs -v`.

[#163539592]

Co-authored-by: Will Murphy <wmurphy@pivotal.io>