Add username and password authentication to login
As part of this work we needed to expose a way for LoginCommand.Execute
to reload the LoginActor's UAAClient so that the Client would have a
non-nil value.
We exposed a new shared function, NewActor, that uses the
return values of shared.NewClients to create a new v2action.Actor; in
the login command, the ActorMaker* types are there so that we can fake
out the behavior of NewActor. Otherwise, we would have to fake out all
of the calls to collaborator objects in shared.NewClients.
This implementation works, but it's not particularly straightforward,
which indicates deeper problems with the factoring of shared.NewClients.

[#163066118]

Co-authored-by: Will Murphy <wmurphy@pivotal.io>