Fix token refresh for "password" grant_type
Previously, in our config.json, an empty grant type would imply
"password". This broke our pipeline when we rewrote this command. In the
rewritten command, we write "password" as the grant type in the
config.json.

This change also stops skipping one of the login tests when running on
windows.

[#163066118]

Co-authored-by: Tom Viehman <tviehman@pivotal.io>