Update delete-orphaned-routes description to clarify deletion scope
It's unclear from the description how wide a `cf delete-orphaned-routes` command will
look to find orphaned routes, e.g. as when run as an admin or as a SpaceDeveloper who
has access to multiple spaces.

This cosmetic update makes it clear that it's only routes which exist in the current space,
not mapped to an app, which will be deleted.