Warn if current CLI is under minimum required version
v3 API currently doesn't report min_cli_version so we added a v2 actor
and corresponding test seam (CheckerMaker), similar to what we did for the
LoginCommand.Actor.

We found it necessary to write two units, one that
tests that the warning is printed to stderr, and one that uses a single
buffer as both stderr and stdout to test the relative position of the
warning in the overall output.

[Finishes #148943711]

Co-authored-by: Alex Shan <ashan@pivotal.io>