Display all integration test failures in PR builder
We needed to change the v6 suite files from symlinks to real files in
order to avoid clobbering the shared output or vice-versa; using
non-linked files allows us to specify a different filename for the
failure summary file.

We also tweaked the task so that it will always try and run all suites
and fail (if necessary) after aggregating all suite failures.

[#163840174]

Co-authored-by: Tom Viehman <tviehman@pivotal.io>