Add env variable to task file
We use the same set of env variables for all the integration tasks in
our pipeline, but not all of the tasks include
CF_INT_IGNORE_API_VERSION_CHECK in their parameters list. If a parameter
is provided in the pipeline config but not expected by the task,
concourse will print a warning, so this change will make those warnings
go away.