Skip test if needed OIDC credentials are not provided
Deploying CF with an OIDC provider configured is a non-standard
configuration (we use a custom ops file to do it) so the integration
tests shouldn't expect users to be able to provide OIDC credentials.

Removed GetOIDCCredentials because it didn't make sense to have two
methods for fetching OIDC creds from the environment.

[Finishes #163136604]