Consolidate PR template
We've noticed that PR submitters have not been filling out the entire template; we've also received some feedback that the template was overly burdensome to fill out because of the volume of questions.

This change attempts to rectify some of the verbosity by collapsing related "why" questions into a single question. We also removed the "Alternate Designs" and "Possible Drawbacks" sections because we feel those questions are better answered by the PR review process.