Revert to storing grant type as empty string in config
We recently changed the format of the config file to store password uaa
grant type as "password" (it was previously stored as "").

Older versions of the CLI have a bug where any string besides "" in the
grant type field causes the CLI to attempt to get a new access token
through the client credentials grant type. That means that if we store
"password" in the config file, users who downgrade would need to
manually delete their config file. Because manually deleting a file is a
bad experience, we will continue to write "" to mean grant type password
in the config file.

[#165201595](https://www.pivotaltracker.com/story/show/165201595)

Co-authored-by: Andrew Crump <acrump@pivotal.io>