Dont list disabled plans in cf marketplace
Currently executing the `cf marketplace` command will give you
a list of services and related to them plans. If you have the admin role
it will list the disabeled plans aswell.

This commit fixes this issue and `cf marketplace` will be showing
only the plans that are available in the current space.

[#165715245]

Co-authored-by: Florent Flament <fflament@pivotal.io>
Co-authored-by: Aarti Kriplani <akriplani@pivotal.io>
Signed-off-by: William Martin <wmartin@pivotal.io>
Co-authored-by: William Martin <wmartin@pivotal.io>