Update v2Actor to pass AccessToken
Recently we changed `cf auth --client-credentials` to not save the
ClientSecret. This renders "refreshing" for client-credentials
impossible.

Streaming logs through the NOAA Client previously relied on refreshing
to get a new token. This change is to update it to pass the current
access token through by default.

A similar change could be made to v3action and v7action. We decided that
was out of scope for our current release patch.

[#165777485](https://www.pivotaltracker.com/story/show/165777485)

Co-authored-by: Brendan Smith <brsmith@pivotal.io>
Co-authored-by: Will Murphy <wmurphy@pivotal.io>