Refactor set-label commands to use a shared api layer
The APIs for updating metadata on API resources in Cloud Controller
follow a consistent interface. This means that apart from the endpoint
URL, the structure of the PATCH requests to these endpoints is uniform
so we should be able to share much of the API layer across resources
when adding/updating labels (and later annotations).

See the following API docs:
https://v3-apidocs.cloudfoundry.org/version/3.71.0/index.html#metadata

This refactor comes from the following discussion in PR #1631:
https://github.com/cloudfoundry/cli/pull/1631#discussion_r280255971

[#165803403](https://www.pivotaltracker.com/story/show/165803403)

Authored-by: Tim Downey <tdowney@pivotal.io>